sum_mt/requirements.txt
absl-py==0.12.0
accelerate==0.11.0
aiohttp==3.8.1
aiosignal==1.2.0
antlr4-python3-runtime==4.8
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
autocommand==2.2.1
backcall==0.2.0
backports.csv==1.0.7
beautifulsoup4==4.11.1
bert-score==0.3.9
bibtexparser==1.4.0
bleach==5.0.1
blis==0.7.8
boto3==1.17.78
botocore==1.20.78
cached-property==1.5.2
cachetools==4.2.2
catalogue==2.0.8
certifi==2022.9.24
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.1.0
checklist==0.0.11
cheroot==8.6.0
CherryPy==18.8.0
click==8.0.1
cmake==3.20.2
colorama==0.4.5
cryptography==39.0.0
cycler==0.10.0
cymem==2.0.6
Cython==0.29.23
dataclasses==0.6
datasets==1.17.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
docopt==0.6.2
editdistance==0.6.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.0/en_core_web_sm-3.4.0-py3-none-any.whl
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.4.0/en_core_web_trf-3.4.0-py3-none-any.whl
entrypoints==0.4
executing==1.1.1
fastjsonschema==2.16.2
fastrlock==0.8
feedparser==6.0.10
filelock==3.0.12
flatbuffers==1.12
frozenlist==1.3.0
fsspec==2022.5.0
future==0.18.2
gast==0.3.3
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
hydra-core==1.0.6
idna==2.10
importlib-metadata==4.0.1
importlib-resources==5.1.4
inflect==6.0.0
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
iso-639==0.4.5
jaraco.classes==3.2.3
jaraco.collections==3.5.2
jaraco.context==4.1.2
jaraco.functools==3.5.2
jaraco.text==3.9.1
jedi==0.18.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.0.1
jsonargparse==3.13.1
jsonlines==2.0.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.21.0
jupyter_client==7.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
keras-nightly==2.5.0.dev2021032900
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
langcodes==3.3.0
latexcodec==2.0.1
lxml==4.9.1
Markdown==3.3.4
MarkupSafe==2.1.1
matplotlib==3.4.2
matplotlib-inline==0.1.6
mistune==2.0.4
more-itertools==9.0.0
mosestokenizer==1.1.0
multidict==6.0.2
multiprocess==0.70.13
munch==2.5.0
murmurhash==1.0.8
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
nltk==3.6.2
notebook==6.5.1
notebook_shim==0.2.0
numpy==1.23.2
oauthlib==3.1.0
omegaconf==2.0.6
openfile==0.0.7
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.5
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.2
patternfork-nosql==3.6
pdfminer.six==20221105
pdfplumber==0.7.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
portalocker==2.0.0
portend==3.1.0
preshed==3.0.7
prometheus-client==0.15.0
prompt-toolkit==3.0.31
protobuf==3.17.0
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==8.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybtex==0.24.0
pycparser==2.20
pydantic==1.9.2
pyDeprecate==0.3.2
pyemd==0.5.1
Pygments==2.13.0
pyinflect==0.5.1
pylatexenc==2.10
pyparsing==2.4.7
-e git+https://github.com/bheinzerling/pyrouge@08e9cc35d713f718a05b02bf3bb2e29947d436ce#egg=pyrouge
pyrsistent==0.18.1
python-dateutil==2.8.2
python-docx==0.8.11
pytorch-lightning==1.6.4
pytorch-nlp==0.5.0
pytorch-pretrained-bert==0.6.2
pytz==2021.1
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
rebiber==1.1.3
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rouge-score==0.1.2
rsa==4.7.2
s3transfer==0.4.2
sacrebleu==2.2.0
sacremoses==0.0.45
scikit-learn==0.24.0
scipy==1.9.1
Send2Trash==1.8.0
sentencepiece==0.1.97
sgmllib3k==1.0.0
six==1.15.0
smart-open==5.2.1
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-alignments==0.8.6
spacy-legacy==3.0.10
spacy-loggers==1.0.3
spacy-transformers==1.1.8
srsly==2.4.4
stack-data==0.5.1
tabulate==0.8.9
tempora==5.0.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==1.1.0
terminado==0.16.0
tf-slim==1.1.0
thinc==8.1.0
threadpoolctl==2.1.0
tinycss2==1.2.1
tokenizers==0.10.2
toolwrapper==2.1.0
torch==1.10.1
torchmetrics==0.8.2
tornado==6.2
tqdm==4.62.1
traitlets==5.5.0
transformers==4.11.3
tsv==1.2
typer==0.4.2
typing_extensions==4.3.0
uctools==1.3.0
unbabel-comet==1.1.2
Unidecode==1.3.6
urllib3==1.26.4
Wand==0.6.11
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.0.1
widgetsnbextension==4.0.3
wrapt==1.12.1
xxhash==3.0.0
yarl==1.7.2
zc.lockfile==2.0
zipp==3.4.1
prettytable==3.6.0
